+++ sim-eth-basic/src/main/java/example/view/ModelViewState.java
@@ -66,6 +66,7 @@ import com.simsilica.mathd.trans.TransitionBuffer;
 import example.ConnectionState;
 import example.GameSessionState;
 import example.Main;
+import example.TimeState;
 import example.net.GameSessionListener;
 import example.net.client.GameSessionClientService;
 
@@ -91,7 +92,7 @@ public class ModelViewState extends BaseAppState {
     private Map<Integer, Spatial> models = new HashMap<>();
     private SafeArrayList<ObjectInfo> objects = new SafeArrayList<>(ObjectInfo.class); 
 
-    private TimeSource timeSource;
+    private TimeState timeState;     
     private SharedObjectUpdater objectUpdater = new SharedObjectUpdater();
 
 
@@ -124,7 +125,11 @@ public class ModelViewState extends BaseAppState {
         // articles at:
         // https://developer.valvesoftware.com/wiki/Source_Multiplayer_Networking
         // https://developer.valvesoftware.com/wiki/Latency_Compensating_Methods_in_Client/Server_In-game_Protocol_Design_and_Optimization
-        this.timeSource = getState(ConnectionState.class).getRemoteTimeSource();
+        //this.timeSource = getState(ConnectionState.class).getRemoteTimeSource();
+        // 
+        // We now grab time from the TimeState which wraps the TimeSource to give
+        // consistent timings over the whole frame
+        timeState = getState(TimeState.class);
         
         // Still need this listener because it's the only way we know things
         // like player name which we might use later.
@@ -194,7 +199,7 @@ public class ModelViewState extends BaseAppState {
         }
  
         // Grab a consistent time for this frame
-        long time = timeSource.getTime();
+        long time = timeState.getTime();
 
         // Update all of the models
         for( ObjectInfo info : objects.getArray() ) {